refactor/driver: refactor entity, enum and db migration
fix/driver: move dto classes
