+++ driver-service/src/main/java/com/artsem/api/driverservice/model/CarClass.java
@@ -1,14 +1,7 @@
 package com.artsem.api.driverservice.model;
 
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
 public enum CarClass {
-    ECONOM("ECONOM"),
-    COMFORT("COMFORT"),
-    BUSINESS("BUSINESS");
-
-    private final String carClass;
+    ECONOM,
+    COMFORT,
+    BUSINESS;
 }
