+++ driver-service/src/main/java/com/artsem/api/driverservice/model/Driver.java
@@ -3,6 +3,8 @@ package com.artsem.api.driverservice.model;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
+import org.hibernate.annotations.CreationTimestamp;
+import org.hibernate.annotations.UpdateTimestamp;
 
 import java.time.LocalDateTime;
 
@@ -33,26 +35,16 @@ public class Driver {
     @Column(name = "is_free")
     private Boolean isFree;
 
+    @CreationTimestamp
     @Column(name = "created_at")
     private LocalDateTime createdAt;
 
+    @UpdateTimestamp
     @Column(name = "updated_at")
     private LocalDateTime updatedAt;
 
     @OneToOne(cascade = CascadeType.PERSIST)
     @JoinColumn(name = "car_id", referencedColumnName = "id")
     private Car car;
-
-    @PrePersist
-    private void initializeTimestamps() {
-        LocalDateTime now = LocalDateTime.now();
-        this.createdAt = now;
-        this.updatedAt = now;
-    }
-
-    @PreUpdate
-    private void updateTimestamp() {
-        this.updatedAt = LocalDateTime.now();
-    }
 
 }
