+++ driver-service/src/main/java/com/artsem/api/driverservice/model/dto/responce/CarResponseDto.java
@@ -1,4 +1,4 @@
-package com.artsem.api.driverservice.model.dto;
+package com.artsem.api.driverservice.model.dto.responce;
 
 import com.artsem.api.driverservice.model.Car;
 import com.artsem.api.driverservice.model.CarClass;
